+++ services/config/src/index.ts
@@ -1,4 +1,5 @@
 export { useConfig } from './useConfig'
+export { useTimeZoneConversion } from './useTimeZoneConversion'
 export { ConfigProvider } from './ConfigProvider'
 
-export type { Config } from './types'
+export type { Config, Version } from './types'