+++ services/offline/src/lib/__tests__/use-cacheable-section.test.tsx
@@ -1,6 +1,5 @@
-import { AlertsProvider } from '@dhis2/app-service-alerts'
 import { renderHook, act } from '@testing-library/react-hooks'
-import React from 'react'
+import React, { FC } from 'react'
 import {
     errorRecordingMock,
     failedMessageRecordingMock,
@@ -29,15 +28,13 @@ afterEach(() => {
 })
 
 it('renders in the default state initially', () => {
-    const { result } = renderHook(() => useCacheableSection('one'), {
-        wrapper: ({ children }) => (
-            <AlertsProvider>
-                <OfflineProvider offlineInterface={mockOfflineInterface}>
-                    {children}
-                </OfflineProvider>
-            </AlertsProvider>
-        ),
-    })
+    const wrapper: FC = ({ children }) => (
+        <OfflineProvider offlineInterface={mockOfflineInterface}>
+            {children}
+        </OfflineProvider>
+    )
+
+    const { result } = renderHook(() => useCacheableSection('one'), { wrapper })
 
     expect(result.current.recordingState).toBe('default')
     expect(result.current.isCached).toBe(false)
@@ -55,17 +52,14 @@ it('handles a successful recording', async (done) => {
                 { sectionId: sectionId, lastUpdated: new Date() },
             ]),
     }
+    const wrapper: FC = ({ children }) => (
+        <OfflineProvider offlineInterface={testOfflineInterface}>
+            {children}
+        </OfflineProvider>
+    )
     const { result, waitFor } = renderHook(
         () => useCacheableSection(sectionId),
-        {
-            wrapper: ({ children }) => (
-                <AlertsProvider>
-                    <OfflineProvider offlineInterface={testOfflineInterface}>
-                        {children}
-                    </OfflineProvider>
-                </AlertsProvider>
-            ),
-        }
+        { wrapper }
     )
 
     const assertRecordingStarted = () => {
@@ -124,15 +118,12 @@ it('handles a recording that encounters an error', async (done) => {
         ...mockOfflineInterface,
         startRecording: errorRecordingMock,
     }
-    const { result } = renderHook(() => useCacheableSection('one'), {
-        wrapper: ({ children }) => (
-            <AlertsProvider>
-                <OfflineProvider offlineInterface={testOfflineInterface}>
-                    {children}
-                </OfflineProvider>
-            </AlertsProvider>
-        ),
-    })
+    const wrapper: FC = ({ children }) => (
+        <OfflineProvider offlineInterface={testOfflineInterface}>
+            {children}
+        </OfflineProvider>
+    )
+    const { result } = renderHook(() => useCacheableSection('one'), { wrapper })
 
     const assertRecordingStarted = () => {
         expect(result.current.recordingState).toBe('recording')
@@ -171,15 +162,12 @@ it('handles an error starting the recording', async () => {
         ...mockOfflineInterface,
         startRecording: failedMessageRecordingMock,
     }
-    const { result } = renderHook(() => useCacheableSection('err'), {
-        wrapper: ({ children }) => (
-            <AlertsProvider>
-                <OfflineProvider offlineInterface={testOfflineInterface}>
-                    {children}
-                </OfflineProvider>
-            </AlertsProvider>
-        ),
-    })
+    const wrapper: FC = ({ children }) => (
+        <OfflineProvider offlineInterface={testOfflineInterface}>
+            {children}
+        </OfflineProvider>
+    )
+    const { result } = renderHook(() => useCacheableSection('err'), { wrapper })
 
     await expect(result.current.startRecording()).rejects.toThrow(
         'Failed message' // from failedMessageRecordingMock
@@ -197,17 +185,14 @@ it('handles remove and updates sections', async () => {
             ])
             .mockResolvedValueOnce([]),
     }
+    const wrapper: FC = ({ children }) => (
+        <OfflineProvider offlineInterface={testOfflineInterface}>
+            {children}
+        </OfflineProvider>
+    )
     const { result, waitFor } = renderHook(
         () => useCacheableSection(sectionId),
-        {
-            wrapper: ({ children }) => (
-                <AlertsProvider>
-                    <OfflineProvider offlineInterface={testOfflineInterface}>
-                        {children}
-                    </OfflineProvider>
-                </AlertsProvider>
-            ),
-        }
+        { wrapper }
     )
 
     // Wait for state to sync with indexedDB
@@ -235,18 +220,14 @@ it('handles a change in ID', async () => {
                 { sectionId: 'id-one', lastUpdated: new Date() },
             ]),
     }
+    const wrapper: FC = ({ children }) => (
+        <OfflineProvider offlineInterface={testOfflineInterface}>
+            {children}
+        </OfflineProvider>
+    )
     const { result, waitFor, rerender } = renderHook(
-        (...args) => useCacheableSection(...args),
-        {
-            wrapper: ({ children }) => (
-                <AlertsProvider>
-                    <OfflineProvider offlineInterface={testOfflineInterface}>
-                        {children}
-                    </OfflineProvider>
-                </AlertsProvider>
-            ),
-            initialProps: 'id-one',
-        }
+        (id: any) => useCacheableSection(id),
+        { wrapper, initialProps: 'id-one' }
     )
 
     // Wait for state to sync with indexedDB